+++ src/routes/blog/index.svelte
@@ -15,7 +15,6 @@
 	import Loading from '../../components/Loading.svelte'
 	import { fade } from 'svelte/transition';
 	import { stores } from "@sapper/app";
-	let { preloading } = stores();
 
 	export let posts;
 
@@ -38,34 +37,30 @@
 	<title>K-NumaTech ブログ</title>
 </svelte:head>
 
-{#if $preloading}
-	<Loading />
-{:else}
-	<div in:fade>
-		{#if !posts.length}
-			<p>該当記事はありませんでした。</p>
-		{/if}
-		{#each posts as post}
-			<!-- TODO: 二回クリックしないと目次が機能しないので#をつける -->
-			<a rel='prefetch' href={`/blog/${post.id}`} class="post w-full mb-8 rounded overflow-hidden shadow flex flex-col md:flex-row hover:shadow-md transition duration-500">
-				<div class="w-full md:w-1/3 h-40 md:h-auto bg-cover" style="background-image: url({coverImage(post.cover_image)})">
+<div in:fade>
+	{#if !posts.length}
+		<p>該当記事はありませんでした。</p>
+	{/if}
+	{#each posts as post}
+		<!-- TODO: 二回クリックしないと目次が機能しないので#をつける -->
+		<a rel='prefetch' href={`/blog/${post.id}`} class="post w-full mb-8 rounded overflow-hidden shadow flex flex-col md:flex-row hover:shadow-md transition duration-500">
+			<div class="w-full md:w-1/3 h-40 md:h-auto bg-cover" style="background-image: url({coverImage(post.cover_image)})">
+			</div>
+			<!-- <img class="w-full" src={post.image.url} alt="thumbnail"> -->
+			<div class="w-full md:w-2/3 flex flex-col">
+				<div class="px-4 md:px-6 py-4">
+					<div class="font-bold text-xl mb-2">{post.title}</div>
+					<p class="truncate text-gray-500 text-sm">{post.body}</p>
 				</div>
-				<!-- <img class="w-full" src={post.image.url} alt="thumbnail"> -->
-				<div class="w-full md:w-2/3 flex flex-col">
-					<div class="px-4 md:px-6 py-4">
-						<div class="font-bold text-xl mb-2">{post.title}</div>
-						<p class="truncate text-gray-500 text-sm">{post.body}</p>
+				<div class="px-4 md:px-6 pt-4 pb-3 mt-auto flex">
+					<div class="flex flex-wrap items-start">
+						{#each post.category.split(',') as category}
+							<span class="inline-block bg-gray-200 rounded-full px-3 py-1 mb-1 text-xs md:text-sm font-semibold text-gray-700 mr-2">{category}</span>
+						{/each}
 					</div>
-					<div class="px-4 md:px-6 pt-4 pb-3 mt-auto flex">
-						<div class="flex flex-wrap items-start">
-							{#each post.category.split(',') as category}
-								<span class="inline-block bg-gray-200 rounded-full px-3 py-1 mb-1 text-xs md:text-sm font-semibold text-gray-700 mr-2">{category}</span>
-							{/each}
-						</div>
-						<div class="ml-auto mb-1 flex flex-shrink-0 self-end items-center text-gray-500 text-sm"><i class="material-icons text-sm mr-1">access_time</i>{post.createdAt.replace(/T.*Z/g, '')}</div>
-					</div>
+					<div class="ml-auto mb-1 flex flex-shrink-0 self-end items-center text-gray-500 text-sm"><i class="material-icons text-sm mr-1">access_time</i>{post.createdAt.replace(/T.*Z/g, '')}</div>
 				</div>
-			</a>
-		{/each}
-	</div>
-{/if}
+			</div>
+		</a>
+	{/each}
+</div>
