+++ src/routes/index.svelte
@@ -1,20 +1,5 @@
 <script>
-	import { onMount } from 'svelte';
 	import { fade } from 'svelte/transition';
-
-	import TopLoadingTitle from '../components/TopLoadingTitle.svelte'
-
-	let isLoading = true;
-
-	function topLoading() {
-		setTimeout(() => {
-			isLoading = false
-		}, 3000);
-	}
-
-	onMount(() => {
-		topLoading()
-	})
 </script>
 
 <svelte:head>
@@ -27,6 +12,3 @@
 <div class="flex mb-4" in:fade>
 
 </div>
-{#if isLoading}
-	<TopLoadingTitle />
-{/if}
