+++ src/stores/topPageLoading.js
@@ -1,0 +1,5 @@
+import { writable } from 'svelte/store';
+
+const topPageLoading = writable(true);
+
+export { topPageLoading }
